Add banner upload and management feature
Introduced BannerService to store and retrieve the current banner image URL. Updated BannerController to support uploading, resizing, and retrieving the banner image. Modified ImageService to ensure banner images are resized to 1200x400 with center-crop. Updated AdminDashboard.html to add a Banner Management section for uploading and previewing the banner image.